Proof of concept syntax plugin
This is a proof of concept of how a plugin could modify the Astro AST. In this case we are supporting backtick code syntax highlighting. Our markdown support compiles backticks to HTML like `<code class="language-astro">....</code>`.

This plugin takes that AST and modifies it so that it uses the Prism astro component.

Downsides:

* Locks us into the AST a bit since we are exposing it here.

Plussides:

* Fits into our pipeline really well. Can see that there are few code changes to make this possible.