Remove `font-weight:400;` from anchor links - they should be the same weight as the text without a link.
If we wanted to, we *could* set the font-weight as `heavier` maybe? Is that an option? So that it just chooses an increase in font-weight for differentiation from text outside of a link.

This was mainly a problem with headings that have links, the difference in font-weight stuck out like a sore thumb to me.