Update deploy script for gitlab pages
As far as I could find, there is not a `build.outDir` inside of `astro.config.mjs`. I was looking at the types here: https://github.com/snowpackjs/astro/blob/main/packages/astro/src/@types/config.ts. Also, when I would attempt to run the command it would not output it to the public directory which makes me think this option was removed at some point, although it may have existed at some point in time.

For this reason, I think it's easier to just add a script to remove all contents of `public` and plop in everything output to the `dist` folder. This should get everything working as one would expect.