Make wording around `client:*` directives more clear
A friend and I both found the wording on this page to be a touch ambiguous. When does the renderer JS download? Does Astro really delay all JS downloading until the component is ready to be hydrated, or just the component JS?

I think these changes make the behavior crystal clear?