fix: don't require 'paths' object in tsconfig to alias paths
In Astro versions < 2.3.0, simply declaring a 'baseUrl' in your tsconfig
would allow for path aliasing. However, commit 8539eb164 added a
requirement that 'paths' also had to exist for the aliases to be
processed. This commit removes that unnecessary requirement so that
aliasing can be achieved solely through defining a 'baseUrl' again.