feat: make shiki highlighting async
When highlighting code via shiki, we default to enabling all available
languages. This results in shiki internally loading each individual
language module, evaluating it, and keeping hold of it in memory.

Instead, we should be able to load the bare minimum and (try) lazily load any
missing languages later on. By doing this, we're not unnecessarily
loading up dozens of large modules and will only load what the user
consumes.

Since traversal of the AST via unist is synchronous, we first do the
traversal to find the nodes and later asynchronously _process_ the nodes
(i.e. execute the highlighter and mutate the AST).