Allow overriding vite mode from astro config
Currently, the mode property of the vite block of user config is ignored. I've updated it to pass this value through to AstroBuilder if it's not defined on inlineConfig. I've also updated some typing for mode to be a plain string, as it's valid to use any arbitrary string for mode for custom environments