Remove @astrojs/lit (#11680)
* Remove @astrojs/lit

This removes the Lit integration as an official integration. The reasons
for doing so in 5.0:

- Only 1% of Astro users use this integration.
- SSR support in Lit is at a lower-level of support due to be a Labs
project: https://lit.dev/docs/libraries/labs/, and has been in this
state since we added support a couple of years ago.
- The maintenance cost of fixing bugs in this integration is too high
given the low usage. Some PRs for upgrading Lit versions have taken
quite a long time. We can't justify the core team being responsible for
this going forward.
- There used to be community contributions to fix bugs but this has
fallen off for various reasons.

Given that, this PR removes the integration as one that is officially
supported by the Astro core team. Interested community members are
encouraged to fork the integration and continue the development in
another repository.

* Remove e2e tests and examples

* Update lockfile

* Remove ssr-lit.test.js

* Remove error for no client entrypoint

* Remove auto selection of the Lit renderer

* Remove lit package.json

* Remove lit fixture