fix test 2
This time, the problem is certain parts of `astro` don't like depending
on `@astrojs/markdown-remark`, because it brings in `node:fs` somehow.
This on its own is probably a problem, but I think I'll work around it
by moving the relevant shared code into
`@astrojs/internal-helpers/remote` so that we can stay node-free.