fix: run fallback view transition for old page only after dispatching `astro:before-swap` event
Makes the native view transitions and the fallback implementation more
consistent with each other.

View transitions can be skipped by hooking onto the `astro:before-swap`
event and calling `.skipTransition()` on it. With native view
transitions, this works as expected. However, the fallback
implementation runs the transition too early — before the
`astro:before-swap` event. This means that the `.skipTransition()` call
would be executed too late, preventing users from skipping the animation
for transitioning the old page out.

This commit makes it so that the out animation for the fallback view
transition implementation only runs after the `astro:before-swap` event
is dispatched, allowing `.skipTransition()` to correctly take effect.