Fix image tests crashing
This prevents the image tests from crashing. The problem is that there
is a circular reference within the Image component. Because the dev:css
plugin was invalidating the image.css during loading it resulted in the
page never finishing, until it ultimately ran out of memory.

The fix here was to move away from a single global dev:css module and
instead create a virtual:astro:dev:css:${route} module for each of
these.

This fixes the problem in that we don't need to invalidate any more,
each page CSS is represented on its own.