Remove non-sensical routing tests for format: file
We previously had tests that allowed you to use `build.format` of
`'file'` and `trailingSlash: 'always'` which of course cannot work,
/foo.html does not have a trailing slash, and web servers would not
respect links with a trailing slash.

Previously this worked for static generation by accident, because did
not construct params from the pathname, but instead just used the return
value of getStaticPaths.

Because the refactor now uses the entire app infrastructure, these tests
failed.

Remove because they are invalid tests.