Fixes for tests in environment branch (#14865)
* Fixes CSS import order and cloudflare

* fix build

* linting obo

* Remove the unused ssr-manifest module

* fix config-vite tests

* fix content-collections.test.js

* fix entry-file-names.test.js

* Fix redirects-i18n test

* Fix server islands tests

If the module with the server island mapping is in a chunk, need to use
a parent-relative import statement to reach the server island module
itself.

* Prevent CSS from being duplicated between SSR and prerender

* Prevent ?raw CSS imports from being added as styles in dev

* Fixes vitest.test.js

Properly handle isDev in the virtual module

* When prerendering, properly handle trailing slash with base

* Prevent empty CSS chunks from being written to disk

* Update core-image-unconventional-settings tests to use environments API

* Prevent adding CSS to pages not belonged to

* Update astro-global.test.js

Prod now matches dev, we default to adding a trailing slash to the
Astro.request.url. Improves an annoying dev/build different in Astro 5.

* Prepend base to propagated styles

Propagated styles get embedded in the JS and rendered at runtime if they
are used on a page. These need to include the base, which this change
makes happen.

* fix the build

* increase the test timeout

* Use the href for saving unix-style paths for image metadata

* use fileURLToNormalizedPath

* PR comments

* fix build

* add comment on what mergeInlineCss does