Remove virtual page module from importing renderers
This had previously existed for reasons I can't even recall. I think it
was related to pages being separate entrypoints but also needing the
renderers array to pass to the App instance.

Since we no longer have this constraint this commit removes them. This
also fixes a bug with Vue that was a circular reference between the Vue
app entrypoint and renderers.