Fix @astrojs/db tests (#14899)
This fixes the Astro DB tests. there were two issues:

1. Previously there was a race condition that never failed. The new code
path must apparently be faster in development mode, as in these tests we
are able to render a page before the seed file runs. Removed the
setTimeout to make it guaranteed to run. This sacrifices dev server
performance.
2. Some tests reused the same fixture. Previously Astro always changed
the bundle hashes so this was safe to do, but now its not, so updated
teh tests to build to different build directories